roon: Fix reports for enable/disable values changes.
The reports were not sent for values which change visibility based on
some switch (enable/disable or certain values). The audio path cache
did not consider this possibility, and while it should be possible to
fix this bug, more bugs of that kind may lurk around in the code. For
this reason, the reports are always computed for the time being, and
they get filtered out by comparing them to the reports last sent.

This is a bit more time and memory consuming, but it is much safer and
the performance his is most probably not noticeable at all. If we need
to improve performance, then the cache can still be fixed. Be advised,
however, that the test suite needs to go much deeper in this case than
it does now.

Change-Id: I1b7bc20fd45ac6b31f9cbc6bbf8da7bc45e590ae